Began Work on Promo Manager
- Implemented stepper functionality for creating a new promo with basic inputs.
The stepper still needs to be fleshed out and the Promo object model still needs to be finalized
for proper binding and form submition. A Lifespan object has been for Date/Time inputs but might
be redefined later.

Other updates:
Minor tweaks to Package Manager components and an update method has been added to the Package Service.