Updates to Promo Manager
- Created all primary steps for promo-wizard
- Created promo service
- Bound data retrieved from serivce to promo preview component
- Updated promo model and created useful enums/classes to accompany it

A json format has been defined for the Promotion Manager and data can now start being passed around
between components as well as fetched using a service. All forms from the promo-wizard store data locally
and are ready to be used to create a proper model. There are still a few conditional inputs that need
to be addressed as well as proper validation for these inputs.