Updates to Promo Manager and Project Tweaks
Promo Manager
- Implemented history component to retrieve and display active/inactive promos as well as add edit and deactivate/activate buttons
- Finished binding all form inputs to model and adjusted steps and layout
- Formatted preview component more and added more info that is bound to the model
- The Promotion model has been tweaked in accordance with some enum changes. The dummy json data has been restructured accrodingly.

Package Manager
- Successfully bound suboption value to view

Project
- Redefined global css variables in styles.scss file and deleted unneccesary scss files in material-angular core files
- Updated some dependencies

Aside from implementing Create and Reset functionality for the promo form and styling everything, the promo manager is largely done.
With the suboption binding issue fixed, the package manager is largely done and just needs styling to be finished off.
Cleanup of the project continues as I discover more unneeded files and adjust the structure.