Implemneted experimental Promimse funcitonality between Carwash Service and subservices
Previously, subservices were not synced with core service when attempting to fetch respective data
and simply plucked a value from the main Carwash data object.
Implementing promises for the the subservice methods that try to fetch respective data has fixed
the problem but introduced other bugs.
A cleaner solution utilizing resolvers is in the works.