Successfully set of Resolve Service for core data
Previous promise method proved to be complicated and introduced strange bugs.
Solution was to move main fetch logic for both the core carwash object and display package items
into a resolver that is utilized only by the components that need it.
This seems to have fixed any async problems observed before and ensures proper load order.